Code Cleanup
Added changes from prod.
