+++ RouteManager.csproj
@@ -91,9 +91,8 @@
     <Compile Include="ModLoader.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="v2\core\AutoEngineer.cs" />
-    <Compile Include="v2\core\ManagedTrains.cs" />
     <Compile Include="v2\core\SettingsManager.cs" />
-    <Compile Include="v2\core\StationManager.cs" />
+    <Compile Include="v2\core\DestinationManager.cs" />
     <Compile Include="v2\core\TrainManager.cs" />
     <Compile Include="v2\dataStructures\LocoTelem.cs" />
     <Compile Include="v2\dataStructures\Settings.cs" />
